Implemented behavior system, Basic AI movement and A start on actual Combat
there is now a behavior system for the enemies as well as the behaviors/behavior class. in addition to that there is also a health component that can take a scriptable object with data as well as unity events for when the a player or enemy takes damage or dies and a basic interface that will allow the player and enemies to do damage to eachother. in the future this interface will include more like the location of the thing dealing damage and other potential things that may be needed to pass through.